requirements.txt
cloudpickle==0.8.0
graphviz==0.13
imageio==2.9.0
matplotlib==3.1.1
neat-python==0.92
numpy
opencv-python
pickleshare==0.7.5
Pillow==8.2.0
scikit-image
scipy==1.5.4
